# RFC: faster implementation of KObj