--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,8 @@ module k8s.io/klog/v2
 
 go 1.13
 
-require github.com/go-logr/logr v1.0.0
+require (
+	github.com/go-logr/logr v1.0.0
+	k8s.io/api v0.22.1 // indirect
+	k8s.io/apimachinery v0.22.1 // indirect
+)