**What this PR does / why we need it**:

`KObj` always does some work (construct `ObjectReference`) which might not be needed later on and also is more expensive to copy.

An alternative implementation (tentatively called `KObj2` at the moment) just stores the interface and uses it later on demand.

**Special notes for your reviewer**:

When testing with `v1.Pod`, the benchmark gives:
```
$ go test -bench=. ./benchmark
goos: linux
goarch: amd64
pkg: k8s.io/klog/v2/benchmark
cpu: Intel(R) Core(TM) i9-7980XE CPU @ 2.60GHz
BenchmarkKObjByPointer-36           	 3013046	       383.2 ns/op
BenchmarkKObj2ByPointer-36          	 3076359	       365.3 ns/op
BenchmarkSkipObjByPointer-36        	 5570828	       234.7 ns/op
BenchmarkSkipObj2ByPointer-36       	 6675914	       178.9 ns/op
BenchmarkDiscardObjByPointer-36     	 5869794	       213.2 ns/op
BenchmarkDiscardObj2ByPointer-36    	 7232146	       163.3 ns/op
PASS
ok  	k8s.io/klog/v2/benchmark	8.829s
```

The commit which adds testing with `v1.Pod` shouldn't be merged because we probably don't want to depend on the k8s API. But it's more realistic than the benchmark with a fake object from the first commit, although that also shows an advantage.

**Release note**:

TBD

```release-note
```