@pohly 2021-09-07
I'm just sharing to document the alternative approach. I'm not sure whether adding it to the klog API now is still worthwhile. It cannot replace the existing `KObj` (API break!), so we would have to carry both.