@neolit123 2021-09-07
> KObj always does some work (construct ObjectReference) which might not be needed later on and also is more expensive to copy.

make sense to me.

leaving review to sig-instrumentation.
/uncc