@serathius 2021-09-07
Recommend using https://github.com/golang/perf/tree/master/cmd/benchstat when analysing performance improvements in golang. This article explains the difference https://about.sourcegraph.com/go/gophercon-2019-optimizing-go-code-without-a-blindfold/#Benchmarks-3-Statistics