@pohly 2021-09-09
"benchstat" is indeed useful. I reorganized the benchmark so that it just has cases with `KObj` which then can be run as-is or with `KObj` replaced by `KObj2`: https://github.com/pohly/klog/commits/testinglogger-kobj2

I also included zapr.

Here the results, including allocs (via `go test -bench=. -count=5 -benchmem ./benchmark/`):
```
name            old time/op    new time/op    delta
String-36          388ns ± 3%     388ns ± 2%     ~     (p=1.000 n=5+5)
Klog-36           5.00µs ± 3%    4.95µs ± 6%     ~     (p=0.389 n=5+5)
KlogSkipped-36     234ns ± 1%     178ns ± 2%  -23.87%  (p=0.008 n=5+5)
Discard-36         204ns ± 3%     148ns ± 1%  -27.09%  (p=0.008 n=5+5)
Zap-36            1.77µs ± 2%    1.61µs ± 1%   -9.29%  (p=0.008 n=5+5)
ZapSkipped-36      245ns ± 2%     191ns ± 4%  -22.19%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
String-36          64.0B ± 0%     64.0B ± 0%     ~     (all equal)
Klog-36             584B ± 0%      568B ± 0%   -2.81%  (p=0.008 n=5+5)
KlogSkipped-36     64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.008 n=5+5)
Discard-36         48.0B ± 0%     32.0B ± 0%  -33.33%  (p=0.008 n=5+5)
Zap-36              193B ± 0%      177B ± 0%   -8.29%  (p=0.008 n=5+5)
ZapSkipped-36      64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
String-36           3.00 ± 0%      3.00 ± 0%     ~     (all equal)
Klog-36             14.0 ± 0%      14.0 ± 0%     ~     (all equal)
KlogSkipped-36      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
Discard-36          2.00 ± 0%      2.00 ± 0%     ~     (all equal)
Zap-36              6.00 ± 0%      6.00 ± 0%     ~     (all equal)
ZapSkipped-36       2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```