@pohly 2021-12-10
/close

I'll come back to this as part of https://github.com/kubernetes/enhancements/pull/3078. I've created a tracking issue at https://github.com/kubernetes/kubernetes/issues/106945